requirements.txt
Flask==3.0.0
numpy>=2.1.0
pandas>=2.2.0
scikit-learn>=1.5.0
google-genai==1.0.0
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
